Add tests, docs, CI and refactor handlers
